Arquivo de configuracao armazenado junto ao evento para facilitar backup.

+++ mkevent.sh
@@ -13,11 +13,16 @@ shift
 
 echo "Creating event: ${codiname}"
 
-"$base/tchelinux-event.py" ${codiname}
+if ! "$base/tchelinux-event.py" ${codiname}
+then
+    echo "Failed to generate hotsite."
+    exit 1
+fi
 
 [ -d ${codiname} ] && rm -rf ${codiname}
-mkdir ${codiname}
+mkdir -p ${codiname}/data
 mv CNAME index.html ${codiname}
 cp -r template/* ${codiname}
 [ -d "images" ] && cp -r images ${codiname} || mkdir ${codinome}/images
+cp data/${codiname}.json ${codiname}/data/${codiname}.json
